Fortify a model into flowSet object

The method provides a universe interface to convert a generic R object into a flowSet useful for ggcyto

fortify_fs(model, data, ...)

# S3 method for flowSet
fortify_fs(model, data, ...)

# S3 method for default
fortify_fs(model, data, ...)

# S3 method for flowFrame
fortify_fs(model, data, ...)

# S3 method for GatingSetList
fortify_fs(model, data, ...)

# S3 method for GatingSet
fortify_fs(model, data, ...)

Arguments

model
flow object(flowFrame or GatingSet) to be converted to flowSet. when it is a GatingSet, it must contain the subset information stored as 'subset' attribute.

data
original dataset, if needed

...
other arguments passed to methods

Value

a flowSet/ncdfFlowSet object

Examples

data(GvHD)
fr <- GvHD[[1]]
fortify_fs(fr)

#> A flowSet with 1 experiments.
#> 
#>   column names:
#>   FSC-H SSC-H FL1-H FL2-H FL3-H FL2-A FL4-H Time

dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
attr(gs, "subset") <- "CD4"
fortify_fs(gs)

#> A cytoset with 1 samples.
#> 
#>   column names:
#>     FSC-A, FSC-H, FSC-W, SSC-A, <B710-A>, <R660-A>, <R780-A>, <V450-A>, <V545-A>, <G560-A>, <G780-A>, Time
#>